+++ tests/sensors/sun.py
@@ -26,16 +26,18 @@ class TestSensorTemperatureDifference(tests.helper.test_case_base.TestCaseBase):
 		tests.helper.oh_item.add_mock_item(HABApp.openhab.items.NumberItem, "H_Unittest_Temperature_1_Unittest_Temperature_2_diff", None)
 		tests.helper.oh_item.add_mock_item(HABApp.openhab.items.NumberItem, "H_Unittest_Temperature_1_Unittest_Temperature_2_diff_filtered", None)
 
-		with unittest.mock.patch("HABApp.openhab.interface.create_item"), unittest.mock.patch("habapp_rules.common.filter.ExponentialFilter"):
+		with unittest.mock.patch("HABApp.openhab.interface.item_exists", return_value=True), unittest.mock.patch("habapp_rules.common.filter.ExponentialFilter"):
 			self._sensor = habapp_rules.sensors.sun.SensorTemperatureDifference("Unittest_Temperature_1", "Unittest_Temperature_2", "Unittest_Output_Temperature", "Unittest_Threshold_Temperature")
 
 	def test_init(self):
 		"""Test __init__."""
 		self.assertEqual(None, self._sensor._threshold)
+		self.assertEqual("H_Unittest_Temperature_1_Unittest_Temperature_2_diff", self._sensor._item_temp_diff.name)
+		self.assertEqual("H_Unittest_Temperature_1_Unittest_Temperature_2_diff_filtered", self._sensor._item_input_filtered.name)
 
 	def test_init_with_fixed_threshold(self):
 		"""Test __init__ with fixed threshold value."""
-		with unittest.mock.patch("HABApp.openhab.interface.create_item"), unittest.mock.patch("habapp_rules.common.filter.ExponentialFilter"):
+		with unittest.mock.patch("HABApp.openhab.interface.item_exists", return_value=True), unittest.mock.patch("habapp_rules.common.filter.ExponentialFilter"):
 			sensor = habapp_rules.sensors.sun.SensorTemperatureDifference("Unittest_Temperature_1", "Unittest_Temperature_2", "Unittest_Output_Temperature", 42)
 		self.assertEqual(42, sensor._threshold)
 
@@ -106,16 +108,17 @@ class TestSensorBrightness(tests.helper.test_case_base.TestCaseBase):
 		tests.helper.oh_item.add_mock_item(HABApp.openhab.items.NumberItem, "Unittest_Threshold_Brightness", None)
 		tests.helper.oh_item.add_mock_item(HABApp.openhab.items.NumberItem, "H_Unittest_Brightness_filtered", None)
 
-		with unittest.mock.patch("HABApp.openhab.interface.create_item"), unittest.mock.patch("habapp_rules.common.filter.ExponentialFilter"):
+		with unittest.mock.patch("HABApp.openhab.interface.item_exists", return_value=True), unittest.mock.patch("habapp_rules.common.filter.ExponentialFilter"):
 			self._sensor = habapp_rules.sensors.sun.SensorBrightness("Unittest_Brightness", "Unittest_Output_Brightness", "Unittest_Threshold_Brightness")
 
 	def test_init(self):
 		"""Test __init__."""
 		self.assertEqual(None, self._sensor._threshold)
+		self.assertEqual("H_Unittest_Brightness_filtered", self._sensor._item_input_filtered.name)
 
 	def test_init_with_fixed_threshold(self):
 		"""Test __init__ with fixed threshold value."""
-		with unittest.mock.patch("HABApp.openhab.interface.create_item"), unittest.mock.patch("habapp_rules.common.filter.ExponentialFilter"):
+		with unittest.mock.patch("HABApp.openhab.interface.item_exists", return_value=True), unittest.mock.patch("habapp_rules.common.filter.ExponentialFilter"):
 			sensor = habapp_rules.sensors.sun.SensorBrightness("Unittest_Brightness", "Unittest_Output_Brightness", 42)
 		self.assertEqual(42, sensor._threshold)